enhance team info
