+++ src/components/beta/team_information_page/team_info_page_body.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, useCallback, useRef } from 'react';
+import { useState } from 'react';
 import Image from 'next/image';
 import { ITeam } from '@/interfaces/team';
 import UploadTeamImageModal from '@/components/beta/team_page/upload_team_image_modal';
@@ -6,51 +6,17 @@ import TeamHeader from '@/components/beta/team_page/team_header';
 import { useTranslation } from 'next-i18next';
 import NoTeamInfo from '@/components/beta/team_information_page/no_team_info';
 import TeamInformation from '@/components/beta/team_information_page/teamInformation';
-import APIHandler from '@/lib/utils/api_handler';
-import { APIName } from '@/constants/api_connection';
-import { useRouter } from 'next/router';
 
 interface TeamPageBodyProps {
   team: ITeam;
+  setTeam: (team: ITeam) => void;
 }
 
-const TeamInformationPageBody = ({ team }: TeamPageBodyProps) => {
+const TeamInformationPageBody = ({ team, setTeam }: TeamPageBodyProps) => {
   const { t } = useTranslation(['team']);
-  const router = useRouter();
-  const { teamId } = router.query; // Info:(20250224 - Anna) 取得網址的 teamId
-  const [teamInfo, setTeamInfo] = useState<ITeam | undefined>();
-  const hasFetched = useRef(false); // Info:(20250226 - Anna) 使用 useRef 避免 API 被執行兩次
+
   const [teamToChangeImage, setTeamToChangeImage] = useState<ITeam | undefined>();
-  const isNoTeamInfo = !teamInfo;
-
-  // Info: (20250226 - Anna) 取得團隊 Info API
-  const { trigger: getTeamInfoByTeamIdAPI } = APIHandler<ITeam>(APIName.GET_TEAM_BY_ID);
-
-  // Info: (20250226 - Anna) 打 API 取得團隊 Info
-  const getTeamInfoByTeamId = useCallback(async () => {
-    if (!teamId || hasFetched.current) return; // Info:(20250226 - Anna) 確保 API 只打一次
-    hasFetched.current = true; // Info:(20250226 - Anna) 標記已執行過 API
-
-    try {
-      const { data: teamInfoData, success } = await getTeamInfoByTeamIdAPI({
-        params: { teamId },
-      });
-      // Info: (20250226 - Anna) 打印 API 回傳的資料（Debug）
-      // eslint-disable-next-line no-console
-      console.log('API 回傳資料:', teamInfoData);
-      if (success && teamInfoData) {
-        setTeamInfo(teamInfoData);
-      }
-    } catch (error) {
-      // Deprecated: (20250226 - Anna) 打印錯誤訊息（Debug）
-      // eslint-disable-next-line no-console
-      console.log('取得團隊資訊失敗');
-    }
-  }, []);
-
-  useEffect(() => {
-    getTeamInfoByTeamId(); // Info: (20250226 - Anna) 在 useEffect 中調用 API
-  }, [teamId]);
+  const isNoTeamInfo = !team;
 
   return (
     <main className="flex flex-col gap-40px">
@@ -67,7 +33,7 @@ const TeamInformationPageBody = ({ team }: TeamPageBodyProps) => {
       {isNoTeamInfo && <NoTeamInfo />}
 
       {/* Info:(20250307 - Anna) 讓子組件能直接更新 teamInfo，團隊修改後 UI 立即更新 */}
-      {!isNoTeamInfo && <TeamInformation teamInfo={teamInfo} setTeamInfo={setTeamInfo} />}
+      {!isNoTeamInfo && <TeamInformation teamInfo={team} setTeamInfo={setTeam} />}
 
       {/* Info: (20250226 - Anna) Modals */}
       {teamToChangeImage && (
